task and auth work fine

+++ app/Http/Controllers/TaskController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
-use App\Task;
+namespace Todo_list\Http\Controllers;
+use Illuminate\Support\Facades\Auth;
+use Todo_list\Task;
 use Illuminate\Http\Request;
 
 class TaskController extends Controller
@@ -13,7 +14,15 @@ class TaskController extends Controller
      */
     public function index()
     {
-        $tasks = Task::orderBy('id', 'DESC')->paginate(5);
+        if (!Auth::check())
+        {
+            return redirect()->guest('login');
+        }
+
+        $user_id = Auth::id();
+
+        $tasks = Task::where('user_id', $user_id)->get();
+
         return view('Task.index', compact('tasks'));
     }
 
@@ -35,6 +44,13 @@ class TaskController extends Controller
      */
     public function store(Request $request)
     {
+        if (!Auth::check())
+        {
+            return redirect()->guest('login');
+        }
+
+        $id = Auth::id();
+
         $this->validate($request, [
             'name' => 'required'
         ]);
@@ -42,7 +58,7 @@ class TaskController extends Controller
         $allFields = $request->all();
         $allFields['start_date'] = date('Y-m-d G:i:s');
         $allFields['end_date'] = date('Y-m-d G:i:s');
-        $allFields['user_id'] = 1;
+        $allFields['user_id'] = $id;
         $allFields['type_id'] = 1;
 
         Task::create($allFields);
@@ -60,6 +76,11 @@ class TaskController extends Controller
      */
     public function show($id)
     {
+        if (!Auth::check())
+        {
+            return redirect()->guest('login');
+        }
+
         $task = Task::find($id);
         return view('Task.show', compact($task));
     }
@@ -72,6 +93,11 @@ class TaskController extends Controller
      */
     public function edit($id)
     {
+        if (!Auth::check())
+        {
+            return redirect()->guest('login');
+        }
+
         $task = Task::find($id);
         return view('Task.edit', compact('task'));
     }
@@ -85,6 +111,13 @@ class TaskController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if (!Auth::check())
+        {
+            return redirect()->guest('login');
+        }
+
+        $user_id = Auth::id();
+
         $this->validate($request, [
             'name' => 'required'
         ]);
@@ -92,7 +125,7 @@ class TaskController extends Controller
         $allFields = $request->all();
         $allFields['start_date'] = date('Y-m-d G:i:s');
         $allFields['end_date'] = date('Y-m-d G:i:s');
-        $allFields['user_id'] = 1;
+        $allFields['user_id'] = $user_id;
         $allFields['type_id'] = 1;
 
         Task::find($id)->update($allFields);
@@ -108,6 +141,11 @@ class TaskController extends Controller
      */
     public function destroy($id)
     {
+        if (!Auth::check())
+        {
+            return redirect()->guest('login');
+        }
+
         Task::find($id)->delete();
         return redirect()->route('task.index')
             ->with('success', 'Task deleted successfully');
